# contracts/openapi/control-plane.v1.yaml
openapi: 3.1.0
info:
  title: mkEngage API
  version: 1.0.0-draft
  license:
    name: Proprietary
    url: https://mkengage.example/terms
  description: |
    Primary external REST contract (§15) served by the control plane. JSON only;
    versioned base path; cursor pagination; idempotency keys on mutations; RFC 9457
    Problem Details errors; correlation IDs; rate-limit headers.

    Draft scope: core Phase-1 resources. Feature modules (campaigns, billing detail,
    Insights queries) extend this contract in their implementation phases.
    SDK clients (packages/api-client and integrations/*) are generated from this file.
servers:
  - url: https://api.mkengage.example/v1
    description: Region-local API endpoint (org home region; global layer routes, ADR-010)
security:
  - bearerAuth: []
  - apiKey: []
...
paths:
  /conversations:
    get:
      tags: [conversations]
      operationId: listConversations
      summary: List conversations (cursor-paginated)
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - name: status
          in: query
          schema: { type: string, enum: [open, pending, closed, all], default: open }
        - name: assigned_agent_id
          in: query
          schema: { type: string, format: uuid }
        - name: department_id
          in: query
          schema: { type: string, format: uuid }
        - name: contact_id
          in: query
          schema: { type: string, format: uuid }
      responses:
        "200":
          description: Page of conversations
          headers:
            X-Request-Id: { $ref: "#/components/headers/XRequestId" }
            X-RateLimit-Limit: { $ref: "#/components/headers/XRateLimitLimit" }
            X-RateLimit-Remaining: { $ref: "#/components/headers/XRateLimitRemaining" }
            X-RateLimit-Reset: { $ref: "#/components/headers/XRateLimitReset" }
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Conversation" }
                  next_cursor: { type: [string, "null"] }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      tags: [conversations]
      operationId: createConversation
      summary: Create a conversation (API channel)
      parameters: [ { $ref: "#/components/parameters/idempotencyKey" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [channel_id]
              properties:
                channel_id: { type: string, format: uuid }
                contact_id: { type: string, format: uuid }
                visitor_id: { type: string, format: uuid }
                department_id: { type: string, format: uuid }
                chatbot_id: { type: string, format: uuid }
                initial_message: { type: string, maxLength: 16000 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /conversations/{conversation_id}:
    parameters: [ { $ref: "#/components/parameters/conversationId" } ]
    get:
      tags: [conversations]
      operationId: getConversation
      summary: Get a conversation
      responses:
        "200":
          description: Conversation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    patch:
      tags: [conversations]
      operationId: updateConversation
      summary: Assign, transfer, or close (administrative mutations — never the message hot path, ADR-002)
      parameters: [ { $ref: "#/components/parameters/idempotencyKey" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status: { type: string, enum: [open, pending, closed] }
                assigned_agent_id: { type: [string, "null"], format: uuid }
                department_id: { type: string, format: uuid }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /conversations/{conversation_id}/messages:
    parameters: [ { $ref: "#/components/parameters/conversationId" } ]
    get:
      tags: [messages]
      operationId: listMessages
      summary: Sequence-ordered message history (also the replay-gap fill, RULES-message-ordering)
      parameters:
        - name: after_sequence
          in: query
          schema: { type: integer, minimum: 0 }
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Messages ordered by sequence_number ascending
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Message" }
                  next_cursor: { type: [string, "null"] }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      tags: [messages]
      operationId: sendMessageRest
      summary: REST fallback send (widget offline fallback, server integrations). Same idempotency semantics as the socket path.
      parameters: [ { $ref: "#/components/parameters/idempotencyKey" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [content_type, body]
              properties:
                content_type: { type: string, enum: [text, attachment, rich] }
                body: { type: string, maxLength: 16000 }
                attachment_ids:
                  type: array
                  maxItems: 10
                  items: { type: string, format: uuid }
      responses:
        "201":
          description: Accepted and persisted (sequence assigned)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Message" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /conversations/{conversation_id}/assign:
    parameters: [ { $ref: "#/components/parameters/conversationId" } ]
    post:
      tags: [conversations]
      operationId: assignConversation
      summary: Assign a conversation (routing v2, §16) — "me" self-assigns the caller, "auto" runs the department strategy, or a specific agent uuid (must be an active department member)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [assignee]
              properties:
                assignee: { type: string, maxLength: 64, description: '"me" | "auto" | agent uuid' }
      responses:
        "200":
          description: Updated conversation
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Conversation" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /insights/overview:
    get:
      tags: [insights]
      operationId: insightsOverview
      summary: Tenant-scoped analytics overview (mkEngage Insights v1). Aggregates are RLS-scoped to the caller's organization; defaults to the trailing 30 days.
      parameters:
        - name: from
          in: query
          schema: { type: string, format: date }
        - name: to
          in: query
          schema: { type: string, format: date }
      responses:
        "200":
          description: Overview metrics
          content:
            application/json:
              schema:
                type: object
                required: [range, conversations, messages, by_department, daily]
                properties:
                  range:
                    type: object
                    properties:
                      from: { type: string, format: date }
                      to: { type: string, format: date }
                  conversations:
                    type: object
                    properties:
                      total: { type: integer }
                      open: { type: integer }
                      pending: { type: integer }
                      closed: { type: integer }
                      resolution_rate: { type: number }
                  messages:
                    type: object
                    properties:
                      total: { type: integer }
                      by_sender: { type: object, additionalProperties: { type: integer } }
                      automation_rate: { type: number }
                  by_department:
                    type: array
                    items:
                      type: object
                      properties:
                        department_name: { type: string }
                        conversations: { type: integer }
                  daily:
                    type: array
                    items:
                      type: object
                      properties:
                        date: { type: string, format: date }
                        conversations: { type: integer }
                        messages: { type: integer }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /me/availability:
    get:
      tags: [conversations]
      operationId: getAvailability
      summary: The caller agent's own availability (routing v2)
      responses:
        "200":
          description: Availability
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Availability" }
        default: { $ref: "#/components/responses/Problem" }
    patch:
      tags: [conversations]
      operationId: setAvailability
      summary: Set the caller agent's availability / concurrency cap. Only `available` agents receive auto-assignments.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/Availability" }
      responses:
        "200":
          description: Updated availability
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Availability" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /conversations/{conversation_id}/notes:
    parameters: [ { $ref: "#/components/parameters/conversationId" } ]
    get:
      tags: [conversations]
      operationId: listNotes
      summary: Internal agent notes (private to the org; never delivered to the visitor, never part of the message sequence)
      responses:
        "200":
          description: Notes, oldest first
          content:
            application/json:
              schema:
                type: object
                required: [data]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Note" } }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      tags: [conversations]
      operationId: addNote
      summary: Add an internal note to a conversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [body]
              properties:
                body: { type: string, maxLength: 8000 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Note" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /conversations/{conversation_id}/attachments:
    parameters: [ { $ref: "#/components/parameters/conversationId" } ]
    post:
      tags: [messages]
      operationId: uploadAttachment
      summary: Upload a chat attachment (§14 — size/type validated, checksummed, scanned async; link it to a message via attachment_ids at send time)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file: { type: string, format: binary }
      responses:
        "201":
          description: Stored with scan_status=pending; scanning runs asynchronously
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Attachment" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /conversations/{conversation_id}/attachments/{attachment_id}/download:
    parameters:
      - { $ref: "#/components/parameters/conversationId" }
      - name: attachment_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [messages]
      operationId: attachmentDownloadUrl
      summary: Mint a short-lived download URL (§14 pre-signed; clean files only — 409 while pending, 410 when quarantined)
      responses:
        "200":
          description: Pre-signed (S3) or signed-route (local) URL, short expiry
          content:
            application/json:
              schema:
                type: object
                required: [url]
                properties:
                  url: { type: string, format: uri }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /contacts:
    get:
      tags: [contacts]
      operationId: listContacts
      summary: List contacts (cursor-paginated)
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
        - name: email
          in: query
          schema: { type: string, format: email }
      responses:
        "200":
          description: Page of contacts
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Contact" }
                  next_cursor: { type: [string, "null"] }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    post:
      tags: [contacts]
      operationId: createContact
      summary: Create a contact
      parameters: [ { $ref: "#/components/parameters/idempotencyKey" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContactInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contact" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /contacts/{contact_id}:
    parameters:
      - name: contact_id
        in: path
        required: true
        schema: { type: string, format: uuid }
    get:
      tags: [contacts]
      operationId: getContact
      summary: Get a contact
      responses:
        "200":
          description: Contact
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contact" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
    patch:
      tags: [contacts]
      operationId: updateContact
      summary: Update a contact
      parameters: [ { $ref: "#/components/parameters/idempotencyKey" } ]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/ContactInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Contact" }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
  /chatbots:
    get:
      tags: [chatbots]
      operationId: listChatbots
      summary: List chatbots
      parameters:
        - $ref: "#/components/parameters/cursor"
        - $ref: "#/components/parameters/limit"
      responses:
        "200":
          description: Page of chatbots
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data:
                    type: array
                    items: { $ref: "#/components/schemas/Chatbot" }
                  next_cursor: { type: [string, "null"] }
        "4XX": { $ref: "#/components/responses/Problem" }
        default: { $ref: "#/components/responses/Problem" }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: First-party session/personal token (Sanctum) for dashboard and user-context calls.
    apiKey:
      type: apiKey
      in: header
      name: Authorization
      description: "Scoped API key: `Authorization: Bearer mk_{env}_{...}` (§15)."
  parameters:
    conversationId:
      name: conversation_id
      in: path
      required: true
      schema: { type: string, format: uuid }
    cursor:
      name: cursor
      in: query
      schema: { type: string }
      description: Opaque cursor from next_cursor. Cursor pagination only — no offsets (§15).
    limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 25 }
    idempotencyKey:
      name: Idempotency-Key
      in: header
      required: true
      schema: { type: string, format: uuid }
      description: Replays within the retention window return the original result (RULES-failure-retry).
  headers:
    XRequestId:
      schema: { type: string, format: uuid }
      description: Correlation ID; echoed from request or generated (ADR-008). Present on every response.
    XRateLimitLimit:
      schema: { type: integer }
      description: Request quota for the current window. Present on every response (§15).
    XRateLimitRemaining:
      schema: { type: integer }
    XRateLimitReset:
      schema: { type: integer }
      description: Unix seconds when the window resets.
  responses:
    Problem:
      description: RFC 9457 Problem Details error (§15). 404 for cross-tenant/nonexistent alike — existence never leaks (RULES-tenant-isolation).
      content:
        application/problem+json:
          schema:
            type: object
            required: [type, title, status]
            properties:
              type: { type: string, format: uri }
              title: { type: string }
              status: { type: integer }
              detail: { type: string, description: Safe detail; never internals or stack traces. }
              correlation_id: { type: string, format: uuid }
              errors:
                type: object
                description: Field-level validation errors.
                additionalProperties:
                  type: array
                  items: { type: string }
  schemas:
...
    Conversation:
      type: object
      required: [conversation_id, organization_id, channel_id, status, created_at, last_sequence]
      properties:
        conversation_id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        channel_id: { type: string, format: uuid }
        status: { type: string, enum: [open, pending, closed] }
        visitor_id: { type: [string, "null"], format: uuid }
        contact_id: { type: [string, "null"], format: uuid }
        assigned_agent_id: { type: [string, "null"], format: uuid }
        assigned_agent_name: { type: [string, "null"], description: Display name of the assigned agent (routing v2). }
        department_id: { type: [string, "null"], format: uuid }
        chatbot_id: { type: [string, "null"], format: uuid }
        last_sequence: { type: integer, minimum: 0, description: Highest assigned sequence number. }
        csat_rating: { type: [integer, "null"], minimum: 1, maximum: 5, description: Post-close visitor satisfaction rating (Phase 23). }
        csat_comment: { type: [string, "null"], maxLength: 1000 }
        created_at: { type: string, format: date-time }
        closed_at: { type: [string, "null"], format: date-time }
    Message:
      type: object
      required: [message_id, conversation_id, sender_type, sender_id, sequence_number, content_type, body, sent_at, lifecycle_state]
      properties:
        message_id: { type: string, format: uuid }
        conversation_id: { type: string, format: uuid }
        channel_id: { type: string, format: uuid }
        sender_type: { type: string, enum: [visitor, contact, agent, chatbot, system] }
        sender_id: { type: string, format: uuid }
        sequence_number: { type: integer, minimum: 1 }
        content_type: { type: string, enum: [text, attachment, rich, system_note] }
        body: { type: string }
        lifecycle_state: { type: string, enum: [persisted, delivered, read, failed, redacted, deleted], description: "§27 lifecycle; created/accepted are pre-persistence and never visible via REST." }
        attachments:
          type: array
          items: { $ref: "#/components/schemas/Attachment" }
        sent_at: { type: string, format: date-time }
    Attachment:
      type: object
      description: Attachment metadata (§14). Download URLs are minted on demand by the download endpoint — never inlined here, so their short expiry (§14) is measured from the click, not from a cached list response.
      required: [attachment_id, file_name, content_type_header, size_bytes, scan_status]
      properties:
        attachment_id: { type: string, format: uuid }
        file_name: { type: string, maxLength: 255 }
        content_type_header: { type: string, description: "Server-side detected content type (finfo), never the client-supplied header." }
        size_bytes: { type: integer, minimum: 0 }
        scan_status:
          type: string
          enum: [pending, clean, quarantined]
          description: Quarantine-first (§14) — files are born pending; only clean files are downloadable.
    Note:
      type: object
      description: Internal agent note — private to the org, never on the visitor surface.
      required: [note_id, conversation_id, author_id, body, created_at]
      properties:
        note_id: { type: string, format: uuid }
        conversation_id: { type: string, format: uuid }
        author_id: { type: string, format: uuid }
        author_name: { type: [string, "null"] }
        body: { type: string }
        created_at: { type: [string, "null"], format: date-time }
    Availability:
      type: object
      description: Agent availability + concurrency cap (routing v2, §16). Only `available` agents receive auto-assignments.
      required: [availability]
      properties:
        availability: { type: string, enum: [available, away] }
        max_open_conversations: { type: [integer, "null"], minimum: 1, description: "Null = unlimited; otherwise the agent is skipped once this many open conversations are assigned." }
    Contact:
      type: object
      required: [contact_id, organization_id, created_at]
      properties:
        contact_id: { type: string, format: uuid }
        organization_id: { type: string, format: uuid }
        email: { type: [string, "null"], format: email }
        name: { type: [string, "null"] }
        phone: { type: [string, "null"] }
        external_id: { type: [string, "null"], description: "Customer-system identifier (signed identity subject, ADR-009)." }
        attributes:
          type: object
          additionalProperties: { type: [string, number, boolean] }
        created_at: { type: string, format: date-time }
    ContactInput:
      type: object
      properties:
        email: { type: [string, "null"], format: email }
        name: { type: [string, "null"] }
        phone: { type: [string, "null"] }
        external_id: { type: [string, "null"] }
        attributes:
          type: object
          additionalProperties: { type: [string, number, boolean] }
    Chatbot:
      type: object
      required: [chatbot_id, name, status]
      properties:
        chatbot_id: { type: string, format: uuid }
        name: { type: string }
        status: { type: string, enum: [draft, active, paused] }
        flow_version: { type: [integer, "null"], description: Active mkEngage Flow definition version (ASSUMPTIONS A1). }
        knowledge_source_ids:
          type: array
          items: { type: string, format: uuid }
...